# app/internal/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// Defines values for ErrorCode.
const (
	BADREQUEST       ErrorCode = "BAD_REQUEST"
	INTERNALERROR    ErrorCode = "INTERNAL_ERROR"
	INVALIDARGUMENT  ErrorCode = "INVALID_ARGUMENT"
	VALIDATIONFAILED ErrorCode = "VALIDATION_FAILED"
	WORDSNOTFOUND    ErrorCode = "WORDS_NOT_FOUND"
)

// Valid indicates whether the value is a known member of the ErrorCode enum.
func (e ErrorCode) Valid() bool {
	switch e {
	case BADREQUEST:
		return true
	case INTERNALERROR:
		return true
	case INVALIDARGUMENT:
		return true
	case VALIDATIONFAILED:
		return true
	case WORDSNOTFOUND:
		return true
	default:
		return false
	}
}

// ErrorCode defines model for ErrorCode.
type ErrorCode string

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code    ErrorCode `json:"code"`
	Details *string   `json:"details,omitempty"`
	Error   string    `json:"error"`
}

// StatusResponse defines model for StatusResponse.
type StatusResponse struct {
	Success bool `json:"success"`
}

// StopWordsRequest defines model for StopWordsRequest.
type StopWordsRequest struct {
	Words []string `json:"words"`
}

// TopQueriesResponse defines model for TopQueriesResponse.
type TopQueriesResponse struct {
	Data []TopQueryItem `json:"data"`

	// UpdatedAt Время последнего перерасчета топа воркером
	UpdatedAt time.Time `json:"updated_at"`
}

// TopQueryItem defines model for TopQueryItem.
type TopQueryItem struct {
	Count int64  `json:"count"`
	Query string `json:"query"`
}

// GetTopQueriesParams defines parameters for GetTopQueries.
type GetTopQueriesParams struct {
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// DeleteStopWordsJSONRequestBody defines body for DeleteStopWords for application/json ContentType.
type DeleteStopWordsJSONRequestBody = StopWordsRequest

// AddStopWordsJSONRequestBody defines body for AddStopWords for application/json ContentType.
type AddStopWordsJSONRequestBody = StopWordsRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Массово удалить слова из динамического стоп-листа
	// (DELETE /api/v1/stop-list)
	DeleteStopWords(w http.ResponseWriter, r *http.Request)
	// Массово добавить слова в динамический стоп-лист
	// (POST /api/v1/stop-list)
	AddStopWords(w http.ResponseWriter, r *http.Request)
	// Получить топ популярных поисковых запросов за последние 5 минут
	// (GET /api/v1/top)
	GetTopQueries(w http.ResponseWriter, r *http.Request, params GetTopQueriesParams)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Массово удалить слова из динамического стоп-листа
// (DELETE /api/v1/stop-list)
func (_ Unimplemented) DeleteStopWords(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Массово добавить слова в динамический стоп-лист
// (POST /api/v1/stop-list)
func (_ Unimplemented) AddStopWords(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Получить топ популярных поисковых запросов за последние 5 минут
// (GET /api/v1/top)
func (_ Unimplemented) GetTopQueries(w http.ResponseWriter, r *http.Request, params GetTopQueriesParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// DeleteStopWords operation middleware
func (siw *ServerInterfaceWrapper) DeleteStopWords(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteStopWords(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AddStopWords operation middleware
func (siw *ServerInterfaceWrapper) AddStopWords(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AddStopWords(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTopQueries operation middleware
func (siw *ServerInterfaceWrapper) GetTopQueries(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTopQueriesParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTopQueries(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/api/v1/stop-list", wrapper.DeleteStopWords)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/v1/stop-list", wrapper.AddStopWords)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1/top", wrapper.GetTopQueries)
	})

	return r
}

type DeleteStopWordsRequestObject struct {
	Body *DeleteStopWordsJSONRequestBody
}

type DeleteStopWordsResponseObject interface {
	VisitDeleteStopWordsResponse(w http.ResponseWriter) error
}

type DeleteStopWords200JSONResponse StatusResponse

func (response DeleteStopWords200JSONResponse) VisitDeleteStopWordsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type DeleteStopWords400JSONResponse ErrorResponse

func (response DeleteStopWords400JSONResponse) VisitDeleteStopWordsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type DeleteStopWords404JSONResponse ErrorResponse

func (response DeleteStopWords404JSONResponse) VisitDeleteStopWordsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type DeleteStopWords500JSONResponse ErrorResponse

func (response DeleteStopWords500JSONResponse) VisitDeleteStopWordsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type AddStopWordsRequestObject struct {
	Body *AddStopWordsJSONRequestBody
}

type AddStopWordsResponseObject interface {
	VisitAddStopWordsResponse(w http.ResponseWriter) error
}

type AddStopWords201JSONResponse StatusResponse

func (response AddStopWords201JSONResponse) VisitAddStopWordsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type AddStopWords400JSONResponse ErrorResponse

func (response AddStopWords400JSONResponse) VisitAddStopWordsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type AddStopWords422JSONResponse ErrorResponse

func (response AddStopWords422JSONResponse) VisitAddStopWordsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(422)

	return json.NewEncoder(w).Encode(response)
}

type AddStopWords500JSONResponse ErrorResponse

func (response AddStopWords500JSONResponse) VisitAddStopWordsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type GetTopQueriesRequestObject struct {
	Params GetTopQueriesParams
}

type GetTopQueriesResponseObject interface {
	VisitGetTopQueriesResponse(w http.ResponseWriter) error
}

type GetTopQueries200JSONResponse TopQueriesResponse

func (response GetTopQueries200JSONResponse) VisitGetTopQueriesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetTopQueries400JSONResponse ErrorResponse

func (response GetTopQueries400JSONResponse) VisitGetTopQueriesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type GetTopQueries500JSONResponse ErrorResponse

func (response GetTopQueries500JSONResponse) VisitGetTopQueriesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Массово удалить слова из динамического стоп-листа
	// (DELETE /api/v1/stop-list)
	DeleteStopWords(ctx context.Context, request DeleteStopWordsRequestObject) (DeleteStopWordsResponseObject, error)
	// Массово добавить слова в динамический стоп-лист
	// (POST /api/v1/stop-list)
	AddStopWords(ctx context.Context, request AddStopWordsRequestObject) (AddStopWordsResponseObject, error)
	// Получить топ популярных поисковых запросов за последние 5 минут
	// (GET /api/v1/top)
	GetTopQueries(ctx context.Context, request GetTopQueriesRequestObject) (GetTopQueriesResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// DeleteStopWords operation middleware
func (sh *strictHandler) DeleteStopWords(w http.ResponseWriter, r *http.Request) {
	var request DeleteStopWordsRequestObject

	var body DeleteStopWordsJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteStopWords(ctx, request.(DeleteStopWordsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteStopWords")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(DeleteStopWordsResponseObject); ok {
		if err := validResponse.VisitDeleteStopWordsResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// AddStopWords operation middleware
func (sh *strictHandler) AddStopWords(w http.ResponseWriter, r *http.Request) {
	var request AddStopWordsRequestObject

	var body AddStopWordsJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.AddStopWords(ctx, request.(AddStopWordsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "AddStopWords")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(AddStopWordsResponseObject); ok {
		if err := validResponse.VisitAddStopWordsResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetTopQueries operation middleware
func (sh *strictHandler) GetTopQueries(w http.ResponseWriter, r *http.Request, params GetTopQueriesParams) {
	var request GetTopQueriesRequestObject

	request.Params = params

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetTopQueries(ctx, request.(GetTopQueriesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetTopQueries")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetTopQueriesResponseObject); ok {
		if err := validResponse.VisitGetTopQueriesResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}